Fixed error handling bug.
Previously, in the event of an error in the profile tasks, we would invoke
the asyncCallback with "null" and {error: error message}. But we also
needed to include a third parameter, browser, since we want to move onto
the next stage in the async-waterfall. Without the third param, we will
essentially stall/spin in the current stage.