Optimized recommendations extraction process.
As with connections, we previously had to trawl through an infinite
scrolling mechanism for profiles with many recommendations.
Now, we are invoking a specific URL instead to get all of the recs in 1
go.
As a part of this process, we need the profile id as well, which is now
being recorded in get-basic-profile-details, the 1st step of the entire
profile waterfall.